Refactoring get_best_league into teamgen
